[BEAM-2206] Move pipelineOptions into options modules
