+++ sdks/python/apache_beam/examples/cookbook/custom_ptransform.py
@@ -28,7 +28,7 @@ import logging
 import apache_beam as beam
 from apache_beam.io import ReadFromText
 from apache_beam.io import WriteToText
-from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.options.pipeline_options import PipelineOptions
 
 
 # pylint doesn't understand our pipeline syntax:
