+++ sdks/python/apache_beam/examples/cookbook/multiple_output_pardo.py
@@ -59,8 +59,8 @@ import apache_beam as beam
 from apache_beam import pvalue
 from apache_beam.io import ReadFromText
 from apache_beam.io import WriteToText
-from apache_beam.utils.pipeline_options import PipelineOptions
-from apache_beam.utils.pipeline_options import SetupOptions
+from apache_beam.options.pipeline_options import PipelineOptions
+from apache_beam.options.pipeline_options import SetupOptions
 
 
 class SplitLinesToWordsFn(beam.DoFn):
