+++ sdks/python/apache_beam/options/__init__.py
@@ -14,8 +14,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-"""For backwards compatibility"""
-
-# pylint: disable=unused-import
-from apache_beam.options.value_provider import *
