+++ sdks/python/apache_beam/runners/dataflow/internal/dependency_test.py
@@ -26,9 +26,9 @@ import unittest
 from apache_beam.io.filesystems import FileSystems
 from apache_beam.runners.dataflow.internal import dependency
 from apache_beam.runners.dataflow.internal import names
-from apache_beam.utils.pipeline_options import GoogleCloudOptions
-from apache_beam.utils.pipeline_options import PipelineOptions
-from apache_beam.utils.pipeline_options import SetupOptions
+from apache_beam.options.pipeline_options import GoogleCloudOptions
+from apache_beam.options.pipeline_options import PipelineOptions
+from apache_beam.options.pipeline_options import SetupOptions
 
 
 class SetupTest(unittest.TestCase):
