+++ sdks/python/apache_beam/runners/dataflow/test_dataflow_runner.py
@@ -18,7 +18,7 @@
 """Wrapper of Beam runners that's built for running and verifying e2e tests."""
 
 from apache_beam.internal import pickler
-from apache_beam.utils.pipeline_options import TestOptions, GoogleCloudOptions
+from apache_beam.options.pipeline_options import TestOptions, GoogleCloudOptions
 from apache_beam.runners.dataflow.dataflow_runner import DataflowRunner
