+++ sdks/python/apache_beam/runners/direct/direct_runner.py
@@ -32,8 +32,8 @@ from apache_beam.runners.runner import PipelineResult
 from apache_beam.runners.runner import PipelineRunner
 from apache_beam.runners.runner import PipelineState
 from apache_beam.runners.runner import PValueCache
-from apache_beam.utils.pipeline_options import DirectOptions
-from apache_beam.utils.value_provider import RuntimeValueProvider
+from apache_beam.options.pipeline_options import DirectOptions
+from apache_beam.options.value_provider import RuntimeValueProvider
 
 
 class DirectRunner(PipelineRunner):
