+++ sdks/python/apache_beam/runners/runner_test.py
@@ -38,7 +38,7 @@ from apache_beam.runners import DirectRunner
 from apache_beam.runners import create_runner
 from apache_beam.transforms.util import assert_that
 from apache_beam.transforms.util import equal_to
-from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.options.pipeline_options import PipelineOptions
 
 
 class RunnerTest(unittest.TestCase):
