+++ sdks/python/apache_beam/test_pipeline_test.py
@@ -25,7 +25,7 @@ from hamcrest.core.assert_that import assert_that as hc_assert_that
 
 from apache_beam.internal import pickler
 from apache_beam.test_pipeline import TestPipeline
-from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.options.pipeline_options import PipelineOptions
 
 
 # A simple matcher that is ued for testing extra options appending.
