+++ sdks/python/apache_beam/typehints/typed_pipeline_test.py
@@ -26,7 +26,7 @@ from apache_beam import typehints
 from apache_beam.test_pipeline import TestPipeline
 from apache_beam.transforms.util import assert_that, equal_to
 from apache_beam.typehints import WithTypeHints
-from apache_beam.utils.pipeline_options import OptionsContext
+from apache_beam.options.pipeline_options import OptionsContext
 
 # These test often construct a pipeline as value | PTransform to test side
 # effects (e.g. errors).
